fix(tests): extract isEstimateExpiringSoon to break import chain
paypal-express.spec.js crashed with 'window is not defined' because importing paypal-context.js transitively loaded checkout-order.js → commerce-api.js → commerce-config.js, which reads window.location at the top level. Extracts isEstimateExpiringSoon into scripts/estimate-token.js so paypal-context.js no longer drags in the commerce stack. checkout-order.js re-exports the function for backward compatibility.
